chore: remove direct uses of Android classes from GeoLocationViewModel (#1074)
Made LocationService interface to provide access to the administrative
area.

Made two implementations of it: one for playstore and one for
f-droid. The f-droid flavored implementation simply throws an exception
on trying to get the administrative area.

The F-droid flavored ViewModel simply hides the location option from the
user on initialization. The Playstore flavored ViewModel works fully.

Made suitable changes to the Koin code.

Since the new code doesn't require Activity, it is no longer passed in
as a parameter in the configure method.

Fixes: #1072